backend/requirements.txt
mem0ai
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
psycopg2-binary
python-dotenv
python-jose
agate
pinecone 
PyPDF2
groq
agno
